+++ src/index.js
@@ -8,10 +8,12 @@ var http               = require('./http.js');
 var offline            = require('./offline.js');
 var localStorageDriver = require('./localStorageDriver.js');
 var cache              = require('./cache.js');
+var transaction    	   = require('./store/store.transaction.js');
 
 module.exports = {
   _stores             : lunarisExports._stores,
   _collection         : collection.collection,
+  _resetTransaction   : transaction._reset,
   _cache              : cache,
   _resetVersionNumber : collection.resetVersionNumber,
   _indexedDB          : localStorageDriver.indexedDB,
@@ -39,8 +41,8 @@ module.exports = {
   getDefaultValue : store.getDefaultValue,
   validate        : store.validate,
   setPagination   : store.setPagination,
-  begin           : hook.begin,
-  commit          : hook.commit,
+  begin           : transaction.begin,
+  commit          : transaction.commit,
 
   OPERATIONS : utils.OPERATIONS,
   constants  : lunarisExports.constants