+++ src/store/store.hook.js
@@ -1,71 +1,7 @@
-var logger           = require('../logger.js');
-var exportsLunaris   = require('../exports.js');
-var storeUtils       = require('./store.utils.js');
-var storeDependecies = exportsLunaris.storeDependencies;
+var logger      = require('../logger.js');
+var transaction = require('./store.transaction.js');
 
-var events                 = {};
-var isTransaction          = false;
-var isCommitingTransaction = false;
-
-/**
- * Begin a transaction
- * @returns {Int} the transaction id
- */
-function begin () {
-  isTransaction = true;
-}
-
-/**
- * Commit a transaction
- * @param {Int} transactionId
- */
-function commit () {
-  isCommitingTransaction = true;
-  var _stores            = Object.keys(events);
-  var _eventByStores     = {};
-
-  for (var i = 0; i < _stores.length; i++) {
-    var _deps = storeDependecies[_stores[i]];
-    var _hasBeenAdded = false;
-
-    for (var j = 0; j < _deps.length; j++) {
-      if (!_eventByStores[_deps[j]]) {
-        _eventByStores[_deps[j]] = [];
-      }
-
-      if (!_hasBeenAdded) {
-        _eventByStores[_deps[j]].push(_stores[i]);
-        _hasBeenAdded = true;
-      }
-    }
-    _hasBeenAdded = false;
-  }
-
-  var _storesToUpdate = Object.keys(_eventByStores);
-  for (i = 0; i < _storesToUpdate.length; i++) {
-    var _storesUpdated = _eventByStores[_storesToUpdate[i]];
-
-    if (_storesUpdated.length) {
-      var _store = storeUtils.getStore(_storesUpdated[_storesUpdated.length - 1]);
-      pushToHandlers(_store, 'filterUpdated');
-    }
-  }
-
-  // Reset
-  isTransaction          = false;
-  isCommitingTransaction = false;
-  events                 = {};
-}
-
-/**
- * Add event to send
- * @param {Stirng} store name
- * @param {String} event event
- */
-function _addEvent (store, event) {
-  events[store] = event;
-}
-
+transaction.registerHookFn(pushToHandlers);
 
 /**
  * Extract the store name and the event from the given string
@@ -174,19 +110,26 @@ function removeHook (hook, handler) {
  * @param {String} hook
  * @param {*} payload
  * @param {Boolean} isMultipleArgsPayload
+ * @param {Int} transactionId
  */
-function pushToHandlers (store, hook, payload, isMultipleArgsPayload) {
+function pushToHandlers (store, hook, payload, isMultipleArgsPayload, transactionId) {
   var _storeHooks = store.hooks[hook];
-  if (!_storeHooks) {
-    return;
-  }
 
-  if (isTransaction && !isCommitingTransaction && hook === 'filterUpdated') {
-    return _addEvent(store.name, hook);
+  if (!_storeHooks) {
+    if (transaction.isCommitingTransaction) {
+      transaction.pipe(store, hook, payload, transactionId);
+    }
+    return;
   }
 
   if (!Array.isArray(payload)) {
     payload = [payload];
+  }
+
+  if (transaction.isCommitingTransaction && hook === 'filterUpdated') {
+    transaction.addUniqueEvent(store.name, hook);
+    transaction.pipe(store, hook, payload, transactionId);
+    return;
   }
 
   for (var i = 0; i < _storeHooks.length; i++) {
@@ -197,10 +140,12 @@ function pushToHandlers (store, hook, payload, isMultipleArgsPayload) {
       _storeHooks[i].apply(null, payload);
     }
   }
+
+  if (transaction.isCommitingTransaction) {
+    transaction.pipe(store, hook, payload, transactionId);
+  }
 }
 
 exports.hook           = registerHook;
 exports.removeHook     = removeHook;
 exports.pushToHandlers = pushToHandlers;
-exports.begin         = begin;
-exports.commit        = commit;