+++ src/store/store.js
@@ -1,20 +1,22 @@
-var lunarisExports     = require('../exports.js');
-var hook               = require('./store.hook.js');
-var utils              = require('../utils.js');
-var storeUtils         = require('./store.utils.js');
-var http               = require('../http.js');
-var logger             = require('../logger.js');
-var cache              = require('../cache.js');
-var md5                = require('../md5.js');
-var url                = require('./store.url.js');
-var template           = require('./store.template.js');
-var collection         = require('./store.collection.js');
-var offline            = require('../offline.js');
-var storeOffline       = require('./store.offline.js');
-var indexedDB          = require('../localStorageDriver.js').indexedDB;
-var OPERATIONS         = utils.OPERATIONS;
-var emptyObject        = {};
-var getRequestQueue    = {};
+var lunarisExports  = require('../exports.js');
+var hook            = require('./store.hook.js');
+var utils           = require('../utils.js');
+var storeUtils      = require('./store.utils.js');
+var http            = require('../http.js');
+var logger          = require('../logger.js');
+var cache           = require('../cache.js');
+var md5             = require('../md5.js');
+var url             = require('./store.url.js');
+var template        = require('./store.template.js');
+var collection      = require('./store.collection.js');
+var offline         = require('../offline.js');
+var storeOffline    = require('./store.offline.js');
+var transaction     = require('./store.transaction.js');
+var indexedDB       = require('../localStorageDriver.js').indexedDB;
+var OPERATIONS      = utils.OPERATIONS;
+var emptyObject     = {};
+var getRequestQueue = {};
+var stores          = [];
 
 lunarisExports._stores.lunarisErrors = {
   name                  : 'lunarisErrors',
@@ -122,17 +124,18 @@ function beforeAction (store, value, isNoValue) {
  * @param {String} event
  * @param {Object/Array} value
  * @param {String} message
+ * @param {Int} transactionId
  */
-function afterAction (store, event, value, message) {
+function afterAction (store, event, value, message, transactionId) {
   var _value = null;
   if (value) {
     _value = utils.cloneAndFreeze(value);
   }
   if (message) {
-    return hook.pushToHandlers(store, event, [_value, message]);
+    return hook.pushToHandlers(store, event, [_value, message], false, transactionId);
   }
 
-  hook.pushToHandlers(store, event, _value, Array.isArray(_value));
+  hook.pushToHandlers(store, event, _value, Array.isArray(_value), transactionId);
 }
 
 /**
@@ -167,16 +170,16 @@ function setLunarisError (storeName, method, request, value, version, err, error
  * @param {Boolean} isMultipleItems
  * @param {Boolean} isUpdate
  * @param {Array} pathParts
+ * @param {Int} transactionId
  * @returns {Int} version
  */
-function _upsertCollection (store, collection, value, version, isMultipleItems, isUpdate, pathParts) {
+function _upsertCollection (store, collection, value, version, isMultipleItems, isUpdate, pathParts, transactionId) {
   var _ids        = [];
   var _inputValue = value;
 
   if (pathParts.length) {
     // set or upddate massOperations rules
     store.massOperations[pathParts.join('.')] = value;
-
     var _data = collection.getAll();
     version   = collection.begin();
     for (var i = 0; i < _data.length; i++) {
@@ -223,9 +226,8 @@ function _upsertCollection (store, collection, value, version, isMultipleItems, 
 
   value = collection.commit(version);
 
-
   cache.invalidate(store.name);
-  afterAction(store, isUpdate ? 'update' : 'insert', value);
+  afterAction(store, isUpdate ? 'update' : 'insert', value, null, transactionId);
   _propagate(store, value, isUpdate ? utils.OPERATIONS.UPDATE : utils.OPERATIONS.INSERT);
   if (isUpdate) {
     _propagateReflexive(store, collection, value,  utils.OPERATIONS.UPDATE);
@@ -259,18 +261,19 @@ function _upsertCollection (store, collection, value, version, isMultipleItems, 
  * @param {*} value
  * @param {Boolean} isMultipleItems
  * @param {Int} version
+ * @param {Int} transactionId
  */
-function _upsertHTTP (method, request, isUpdate, store, collection, cache, value, isMultipleItems, version) {
+function _upsertHTTP (method, request, isUpdate, store, collection, cache, value, isMultipleItems, version, transactionId) {
   http.request(method, request, value, function (err, data) {
     if (err) {
       var _error = template.getError(err, store, method, false);
       setLunarisError(store.name, method, request, value, version, err, _error);
       logger.warn(['lunaris.' + (isUpdate ? 'update' : 'insert') + '@' + store.name], err);
-      return hook.pushToHandlers(store, 'errorHttp', [_error, utils.cloneAndFreeze(value)]);
+      return hook.pushToHandlers(store, 'errorHttp', [_error, utils.cloneAndFreeze(value)], false, transactionId);
     }
 
     if (method === OPERATIONS.PATCH) {
-      afterAction(store, 'patched');
+      afterAction(store, 'patched', null, null, transactionId);
       return;
     }
 
@@ -318,10 +321,10 @@ function _upsertHTTP (method, request, isUpdate, store, collection, cache, value
       return;
     }
 
-    afterAction(store, 'update', value);
-    afterAction(store,  isUpdate ? 'updated' : 'inserted', value, template.getSuccess(null, store, method, false));
+    afterAction(store, 'update', value, null, transactionId);
+    afterAction(store,  isUpdate ? 'updated' : 'inserted', value, template.getSuccess(null, store, method, false), transactionId);
     if (store.isFilter) {
-      hook.pushToHandlers(store, 'filterUpdated');
+      hook.pushToHandlers(store, 'filterUpdated', null, false, transactionId);
     }
     _propagate(store, value, utils.OPERATIONS.UPDATE);
     _propagateReflexive(store, collection, value, utils.OPERATIONS.UPDATE);
@@ -339,11 +342,11 @@ function _upsertHTTP (method, request, isUpdate, store, collection, cache, value
  * @param {Boolean} isUpdate
  * @param {Object} retryOptions
  */
-function _upsert (store, collection, pathParts, value, isLocal, isUpdate, retryOptions) {
+function _upsert (store, collection, pathParts, value, isLocal, isUpdate, retryOptions, transactionId) {
   var _isMultipleItems = Array.isArray(value);
   var _version;
   if (!retryOptions) {
-    var _res = _upsertCollection(store, collection, value, _version, _isMultipleItems, isUpdate, pathParts);
+    var _res = _upsertCollection(store, collection, value, _version, _isMultipleItems, isUpdate, pathParts, transactionId);
     _version = _res.version;
     value    = _res.value;
   }
@@ -351,9 +354,10 @@ function _upsert (store, collection, pathParts, value, isLocal, isUpdate, retryO
     _version = retryOptions.version;
   }
 
+
   if (store.isLocal || isLocal) {
     if (store.isFilter) {
-      hook.pushToHandlers(store, 'filterUpdated');
+      hook.pushToHandlers(store, 'filterUpdated', null, false, transactionId);
     }
     return;
   }
@@ -385,7 +389,7 @@ function _upsert (store, collection, pathParts, value, isLocal, isUpdate, retryO
     return;
   }
 
-  _upsertHTTP(_method, _request, isUpdate, store, collection, cache, value, _isMultipleItems, _version);
+  _upsertHTTP(_method, _request, isUpdate, store, collection, cache, value, _isMultipleItems, _version, transactionId);
 }
 
 /**
@@ -433,8 +437,9 @@ function _transformGetCache (collection, values) {
  */
 function _get (store, primaryKeyValue, retryOptions, callback) {
   try {
-    var _options = beforeAction(store, null, true);
-    var _request = '/';
+    var _options       = beforeAction(store, null, true);
+    var _request       = '/';
+    var _transactionId = transaction.getCurrentTransactionId();
 
     if (!retryOptions) {
       _request = url.create(_options.store, 'GET', primaryKeyValue);
@@ -446,10 +451,10 @@ function _get (store, primaryKeyValue, retryOptions, callback) {
 
       if (_cacheValues) {
         if (typeof _cacheValues === 'object') {
-          afterAction(_options.store, 'get', _transformGetCache(_options.collection, _cacheValues));
+          afterAction(_options.store, 'get', _transformGetCache(_options.collection, _cacheValues), null, _transactionId);
           return callback(store);
         }
-        afterAction(_options.store, 'get', []);
+        afterAction(_options.store, 'get', [], null, _transactionId);
         return callback(store);
       }
 
@@ -459,9 +464,9 @@ function _get (store, primaryKeyValue, retryOptions, callback) {
           _options.collection,
           _request
         );
-        afterAction(_options.store, 'get', _res);
+        afterAction(_options.store, 'get', _res, null, _transactionId);
         if (_options.store.isFilter && ((_options.store.isStoreObject && _res) || (!_options.store.isStoreObject && _res.length))) {
-          hook.pushToHandlers(_options.store, 'filterUpdated');
+          hook.pushToHandlers(_options.store, 'filterUpdated', false, null, _transactionId);
         }
         storeUtils.saveState(_options.store, _options.collection, _options.cache);
         return callback(store);
@@ -478,7 +483,7 @@ function _get (store, primaryKeyValue, retryOptions, callback) {
         var _error = template.getError(err, _options.store, 'GET', true);
         setLunarisError(_options.store.name, 'GET', _request, null, null, err, _error);
         logger.warn(['lunaris.get' + store], err);
-        hook.pushToHandlers(_options.store, 'errorHttp', _error);
+        hook.pushToHandlers(_options.store, 'errorHttp', _error, false, _transactionId);
         return callback(store);
       }
 
@@ -509,7 +514,7 @@ function _get (store, primaryKeyValue, retryOptions, callback) {
 
       data = _options.collection.commit(_version);
 
-      afterAction(_options.store, 'get', data);
+      afterAction(_options.store, 'get', data, null, _transactionId);
       _propagate(_options.store, data, utils.OPERATIONS.INSERT);
       if (_options.store.isFilter) {
         hook.pushToHandlers(_options.store, 'filterUpdated');
@@ -556,6 +561,7 @@ function upsert (store, value, isLocal, retryOptions) {
       value = retryOptions.data;
     }
 
+
     var _options = beforeAction(store, value);
     if ((Array.isArray(value) && value[0]._id) || value._id) {
       _isUpdate = true;
@@ -564,17 +570,29 @@ function upsert (store, value, isLocal, retryOptions) {
       _isUpdate = false;
     }
 
+    if (transaction.isTransaction && !transaction.isCommitingTransaction) {
+      return transaction.addAction({
+        id        : transaction.getCurrentTransactionId(),
+        store     : _options.store.name,
+        operation : _isUpdate ? OPERATIONS.UPDATE : OPERATIONS.INSERT,
+        handler   : upsert,
+        arguments : arguments,
+        rollback  : _isUpdate ? upsert : deleteStore
+      });
+    }
+    var _transactionId = transaction.getCurrentTransactionId();
+
     if (_options.store.validateFn && !_storeParts.length) {
       return validate(store, _options.value, _isUpdate, function (res) {
         if (!res) {
           return;
         }
 
-        _upsert(_options.store, _options.collection, _storeParts, _options.value, isLocal, _isUpdate, retryOptions);
+        _upsert(_options.store, _options.collection, _storeParts, _options.value, isLocal, _isUpdate, retryOptions, _transactionId);
       }, _eventName);
     }
 
-    _upsert(_options.store, _options.collection, _storeParts, _options.value, isLocal, _isUpdate, retryOptions);
+    _upsert(_options.store, _options.collection, _storeParts, _options.value, isLocal, _isUpdate, retryOptions, _transactionId);
   }
   catch (e) {
     logger.warn([_eventName], e);
@@ -586,9 +604,10 @@ function upsert (store, value, isLocal, retryOptions) {
  * @param {Object} store
  * @param {Object} collection
  * @param {*} value
- * @param {Boolean} isError
+ * @param {Boolean} isLocal
+ * @param {Int} transactionId
  */
-function _delete (store, collection, value, isLocal) {
+function _delete (store, collection, value, isLocal, transactionId) {
   var _version = collection.begin();
   collection.remove(value, _version, !isLocal);
   value = collection.commit(_version);
@@ -596,7 +615,7 @@ function _delete (store, collection, value, isLocal) {
   if (isLocal && ((!_isArray && !value) || (_isArray && !value.length))) {
     throw new Error('You cannot delete a value not in the store!');
   }
-  afterAction(store, 'delete', value);
+  afterAction(store, 'delete', value, null, transactionId);
   _propagate(store, value, utils.OPERATIONS.DELETE);
   _propagateReflexive(store, collection, value, utils.OPERATIONS.DELETE);
 
@@ -626,10 +645,21 @@ function deleteStore (store, value, retryOptions, isLocal) {
       value = retryOptions.data;
     }
     var _options = beforeAction(store, value);
+    if (transaction.isTransaction && !transaction.isCommitingTransaction) {
+      return transaction.addAction({
+        id        : transaction.getCurrentTransactionId(),
+        store     : _options.store.name,
+        operation : OPERATIONS.DELETE,
+        handler   : deleteStore,
+        arguments : arguments,
+        rollback  : upsert
+      });
+    }
+    var _transactionId = transaction.getCurrentTransactionId();
 
     var _version;
     if (!retryOptions) {
-      var _res = _delete(_options.store, _options.collection, value, true);
+      var _res = _delete(_options.store, _options.collection, value, true, _transactionId);
       _version = _res[0];
       value    = _res[1];
     }
@@ -663,11 +693,14 @@ function deleteStore (store, value, retryOptions, isLocal) {
         var _error = template.getError(err, _options.store, 'DELETE', false);
         setLunarisError(_options.store.name, 'DELETE', _request, value, _version, err, _error);
         logger.warn(['lunaris.delete' + store], err);
-        return hook.pushToHandlers(_options.store, 'errorHttp', [_error, value]);
+        return hook.pushToHandlers(_options.store, 'errorHttp', [_error, value], false, _transactionId);
       }
 
-      value = _delete(_options.store, _options.collection, data);
-      afterAction(_options.store, 'deleted', data, template.getSuccess(null, _options.store, 'DELETE', false));
+      _res = _delete(_options.store, _options.collection, data, false, _transactionId)[1];
+      if (_res) {
+        value = _res;
+      }
+      afterAction(_options.store, 'deleted', value, template.getSuccess(null, _options.store, 'DELETE', false), _transactionId);
     });
   }
   catch (e) {
@@ -700,7 +733,7 @@ function setPagination (store, page, limit) {
  * @param {String} store
  * @param {Boolean} isSilent
  */
-function clear (store, isSilent) {
+function _clear (store, isSilent) {
   try {
     var _options = beforeAction(store, null, true);
 
@@ -723,6 +756,35 @@ function clear (store, isSilent) {
 }
 
 /**
+ * Clear the store collection
+ * @param {String} store
+ * @param {Boolean} isSilent
+ */
+function clear (store, isSilent) {
+  // It is a regex, we must find the stores !
+  if (/\*$/.test(store)) {
+    if (!/^@/.test(store)) {
+      return logger.warn(['lunaris.clear'], new Error('The store key must begin by \'@\''));
+    }
+    if (!stores.length) {
+      stores = Object.keys(lunarisExports._stores);
+    }
+
+    var _keyLength = store.length - 2;
+    var _key       = store.slice(1, _keyLength + 1);
+    for (var i = 0; i < stores.length; i++) {
+      if (stores[i].slice(0, _keyLength) === _key) {
+        _clear('@' + stores[i], isSilent);
+      }
+    }
+
+    return;
+  }
+
+  _clear(store, isSilent);
+}
+
+/**
  * Get values
  * @param {String} store
  * @param {*} primaryKeyValue
@@ -733,6 +795,17 @@ function clear (store, isSilent) {
 function get (store, primaryKeyValue, retryOptions) {
   if (!getRequestQueue[store]) {
     getRequestQueue[store] = [];
+  }
+
+  if (transaction.isTransaction && !transaction.isCommitingTransaction) {
+    return transaction.addAction({
+      id        : transaction.getCurrentTransactionId(),
+      store     : store.replace('@', ''),
+      operation : OPERATIONS.LIST,
+      handler   : _get,
+      arguments : [store, primaryKeyValue, retryOptions, function () {}],
+      rollback  : null
+    });
   }
 
   getRequestQueue[store].push([store, primaryKeyValue, retryOptions]);
@@ -878,18 +951,15 @@ function validate (store, value, isUpdate, callback, eventName) {
   }
 }
 
-exports.get               = get;
-exports.getOne            = getOne;
-exports.insert            = upsert;
-// exports.insertFiltered = upsertFiltered;
-exports.update            = upsert;
-// exports.updateFiltered = upsertFiltered;
-exports.upsert            = upsert;
-exports.delete            = deleteStore;
-exports.clear             = clear;
-exports.retry             = retry;
-exports.rollback          = rollback;
-exports.getDefaultValue   = getDefaultValue;
-exports.validate          = validate;
-exports.setPagination     = setPagination;
-// exports.deleteFiltered = deleteFiltered;
+exports.get             = get;
+exports.getOne          = getOne;
+exports.insert          = upsert;
+exports.update          = upsert;
+exports.upsert          = upsert;
+exports.delete          = deleteStore;
+exports.clear           = clear;
+exports.retry           = retry;
+exports.rollback        = rollback;
+exports.getDefaultValue = getDefaultValue;
+exports.validate        = validate;
+exports.setPagination   = setPagination;