+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Lunaris
 
-## 1.7.1
+## 1.10.0
+- Features:
+- Fix:
+
+## 1.9.0
+*2019-04-01*
+- Features:
+  - Add syntax to clear multiple stores at the same time. Use `lunaris.clear('@myCommonStore*')`,
+- Fix:
+  - Fix store dependencies graph.
+
+## 1.8.0
+*2019-03-04*
 - Features:
   - Do not compress the body of HTTP requests in development mode.
   - Validation function supports undefined values for optional attributes.
   - Add header `Content-Version` with value equals to `2`.
   - Allow multiple stores definition in one file. In module.exports, define an array of store objet instead of a store object `module.exports = [ {}, {}, ... ]`.
+  - Begin/Commit supports lunaris.get/insert/update/upsert/delete with local store or/and default store. The transaction ensure that each action is sequential.
 - Fix:
   - For GET HTTP request, if the cache value was an object, an empty array whould be returned.
   - Fix removeHook. When removing hooks, the function was looking at the same function definition and not the reference.